Compute GDSC signatures using HUGO gene symbols
Convert probe IDs to HUGO symbols and collapse by  the `max(absolute(value))` 
being `value` either `t` for `classic` mode or `logFC` for `log` mode. Then,
create gmts with UP/DOWN signatures if the number of genes that pass the filters
is at least 15. Save the signatures with the name of the common drug + GDSC ID.